Full executable build of cleanmod
This branch differs from cleanmod in that it includes the BouncyCastle
and ThoughtWorks.QRCode DLLs at the top level to facilitate builds, and
includes copies of the bank-note png files in the bin\Debug and
bin\Release subdirectories to allow the executables to find them.  And
the Release build is built.